Modules roughly complete
