+++ ethsigner/commandline/build.gradle
@@ -1,20 +1,6 @@
-/*
- * Copyright 2018 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-apply plugin: 'java-library'
 
 jar {
-  baseName 'ethsigner-core'
+  baseName 'ethsigner-commandline'
   manifest {
     attributes(
       'Specification-Title': baseName,
@@ -27,10 +13,11 @@ jar {
 
 
 dependencies {
-  1
 
   implementation project(':ethsigner:signing-api')
+  implementation project(':ethsigner:core')
 
+  implementation 'info.picocli:picocli'
   implementation 'com.google.guava:guava'
 
   implementation 'org.apache.logging.log4j:log4j-api'
