+++ ethsigner/commandline/src/main/java/tech/pegasys/ethsigner/ApplicationInfo.java
