+++ ethsigner/commandline/src/main/java/tech/pegasys/ethsigner/PlatformDetector.java
