+++ ethsigner/commandline/src/main/java/tech/pegasys/ethsigner/VersionInfo.java
