+++ ethsigner/commandline/src/main/java/tech/pegasys/ethsigner/VersionProvider.java
