--- ethsigner/hashicorpsigner/build.gradle
@@ -1,59 +1,0 @@
-/*
- * Copyright 2018 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-apply plugin: 'java-library'
-
-jar {
-  baseName 'ethsigner-core'
-  manifest {
-    attributes(
-      'Specification-Title': baseName,
-      'Specification-Version': project.version,
-      'Implementation-Title': baseName,
-      'Implementation-Version': calculateVersion()
-      )
-  }
-}
-
-
-dependencies {
-
-  implementation project(':ethsigner:subcommandapi')
-  implementation project(':ethsigner:signingapi')
-  implementation project(':ethsigner:core')
-
-  implementation 'info.picocli:picocli'
-  implementation 'com.google.guava:guava'
-
-  implementation 'org.apache.logging.log4j:log4j-api'
-  runtime 'org.apache.logging.log4j:log4j-core'
-  runtime 'org.apache.logging.log4j:log4j-slf4j-impl'
-
-  implementation 'org.web3j:core'
-  implementation 'org.web3j:crypto'
-  implementation 'org.web3j:eea'
-
-  implementation 'io.vertx:vertx-core'
-  implementation 'io.vertx:vertx-web'
-  implementation 'io.vertx:vertx-web-client'
-
-  testCompile 'io.vertx:vertx-codegen'
-  testImplementation 'junit:junit'
-  testImplementation 'org.assertj:assertj-core'
-  testImplementation 'org.mockito:mockito-core'
-
-  integrationTestImplementation 'io.rest-assured:rest-assured'
-  integrationTestImplementation 'org.assertj:assertj-core'
-  integrationTestImplementation 'org.mock-server:mockserver-netty'
-  integrationTestImplementation 'org.mockito:mockito-core'
-}
