+++ ethsigner/signers/file-based/build.gradle
@@ -11,9 +11,10 @@
  * specific language governing permissions and limitations under the License.
  */
 
+apply plugin: 'java-library'
 
 jar {
-  baseName 'ethsigner'
+  baseName 'ethsigner-signer-file-based'
   manifest {
     attributes(
       'Specification-Title': baseName,
@@ -27,28 +28,31 @@ jar {
 
 dependencies {
 
-  implementation project(':ethsigner:core')
-  implementation project(':ethsigner:signing-api')
-  implementation project(':ethsigner:signers:hashicorp')
-  implementation project(':ethsigner:signers:file-based')
   implementation project(':ethsigner:commandline')
-
+  implementation project(':ethsigner:signing-api')
 
   implementation 'info.picocli:picocli'
-
   implementation 'com.google.guava:guava'
 
   implementation 'org.apache.logging.log4j:log4j-api'
   runtime 'org.apache.logging.log4j:log4j-core'
   runtime 'org.apache.logging.log4j:log4j-slf4j-impl'
 
-  implementation 'io.vertx:vertx-core'
+  implementation 'org.web3j:core'
+  implementation 'org.web3j:crypto'
+  implementation 'org.web3j:eea'
 
+  implementation 'io.vertx:vertx-core'
+  implementation 'io.vertx:vertx-web'
+  implementation 'io.vertx:vertx-web-client'
 
   testCompile 'io.vertx:vertx-codegen'
-
-  testImplementation project(':ethsigner:core')
   testImplementation 'junit:junit'
   testImplementation 'org.assertj:assertj-core'
   testImplementation 'org.mockito:mockito-core'
+
+  integrationTestImplementation 'io.rest-assured:rest-assured'
+  integrationTestImplementation 'org.assertj:assertj-core'
+  integrationTestImplementation 'org.mock-server:mockserver-netty'
+  integrationTestImplementation 'org.mockito:mockito-core'
 }
