--- ethsigner/subcommandapi/build.gradle
@@ -1,31 +1,0 @@
-
-dependencies {
-
-    implementation project(':ethsigner:signingapi')
-    implementation project(':ethsigner:core')
-
-    implementation 'info.picocli:picocli'
-    implementation 'com.google.guava:guava'
-
-    implementation 'org.apache.logging.log4j:log4j-api'
-    runtime 'org.apache.logging.log4j:log4j-core'
-    runtime 'org.apache.logging.log4j:log4j-slf4j-impl'
-
-    implementation 'org.web3j:core'
-    implementation 'org.web3j:crypto'
-    implementation 'org.web3j:eea'
-
-    implementation 'io.vertx:vertx-core'
-    implementation 'io.vertx:vertx-web'
-    implementation 'io.vertx:vertx-web-client'
-
-    testCompile 'io.vertx:vertx-codegen'
-    testImplementation 'junit:junit'
-    testImplementation 'org.assertj:assertj-core'
-    testImplementation 'org.mockito:mockito-core'
-
-    integrationTestImplementation 'io.rest-assured:rest-assured'
-    integrationTestImplementation 'org.assertj:assertj-core'
-    integrationTestImplementation 'org.mock-server:mockserver-netty'
-    integrationTestImplementation 'org.mockito:mockito-core'
-}
